+++ src/app/components/core/filter/filter-list/filter-list.component.css
